+++ seqkit/cmd/util.go
@@ -23,6 +23,7 @@ package cmd
 import (
 	"fmt"
 	"io"
+	"math/rand"
 	"os"
 	"os/exec"
 	"regexp"
@@ -543,4 +544,15 @@ func checkFileFormat(format string) {
 	default:
 		log.Fatal("Invalid format specified:", format)
 	}
+}
+
+func randomFloat64Generator(seed int64) chan float64 {
+	rand.Seed(seed)
+	ch := make(chan float64, 1024)
+	go func() {
+		for {
+			ch <- rand.Float64()
+		}
+	}()
+	return ch
 }