+++ CHANGELOG.md
@@ -4,6 +4,7 @@
     - `seqkit grep`: allowing empty pattern files.
     - `seqkit tab2fx`: fix bug for very long sequences. [#214](https://github.com/shenwei356/seqkit/issues/214)
     - `seqkit fish`: fix range check. [#213](https://github.com/shenwei356/seqkit/issues/213)
+    - `seqkit sample`: slightly faster.
 - [SeqKit v0.16.1](https://github.com/shenwei356/seqkit/releases/tag/v0.16.1)
 [![Github Releases (by Release)](https://img.shields.io/github/downloads/shenwei356/seqkit/v0.16.1/total.svg)](https://github.com/shenwei356/seqkit/releases/tag/v0.16.1)
     - `seqkit shuffle --two-pass`: fix bug introduced in [#173](https://github.com/shenwei356/seqkit/issues/173) . [#209](https://github.com/shenwei356/seqkit/issues/209)